refactored tests to remove duplicated code; refactored test cases to multiple files to improve readability
