+++ package.json
@@ -28,9 +28,9 @@
   "preferGlobal": true,
   "scripts": {
     "dev-install": "tsd install",
-    "test": "./node_modules/mocha/bin/_mocha -R spec ./test/*",
-    "test-coverage": "./node_modules/istanbul/lib/cli.js cover ./node_modules/mocha/bin/_mocha -- -R min ./test/*",
-    "test-travis": "./node_modules/istanbul/lib/cli.js cover ./node_modules/mocha/bin/_mocha -- -R spec ./test/*"
+    "test": "./node_modules/mocha/bin/_mocha -R dot test/**",
+    "test-coverage": "./node_modules/istanbul/lib/cli.js cover ./node_modules/mocha/bin/_mocha -- -R min test/**",
+    "test-travis": "./node_modules/istanbul/lib/cli.js cover ./node_modules/mocha/bin/_mocha -- -R spec test/**"
   },
   "repository": {
     "type": "git",
